@@ -1,10 +1,13 @@
 import os
+from collections.abc import Callable
 from enum import Enum
 from pathlib import Path
-from typing import Callable, Dict
 
 import aenum
+import airflow
 from packaging.version import Version
+
+AIRFLOW_VERSION = Version(airflow.__version__)
 
 BIGQUERY_PROFILE_TYPE = "bigquery"
 DBT_PROFILE_PATH = Path(os.path.expanduser("~")).joinpath(".dbt/profiles.yml")
@@ -55,7 +58,7 @@ def _default_wasb_conn() -> str:
     return WasbHook.default_conn_name  # type: ignore[no-any-return]
 
 
-FILE_SCHEME_AIRFLOW_DEFAULT_CONN_ID_MAP: Dict[str, Callable[[], str]] = {
+FILE_SCHEME_AIRFLOW_DEFAULT_CONN_ID_MAP: dict[str, Callable[[], str]] = {
     "s3": _default_s3_conn,
     "gs": _default_gcs_conn,
     "adl": _default_wasb_conn,
@@ -105,6 +108,7 @@ class ExecutionMode(Enum):
     VIRTUALENV = "virtualenv"
     AZURE_CONTAINER_INSTANCE = "azure_container_instance"
     GCP_CLOUD_RUN_JOB = "gcp_cloud_run_job"
+    WATCHER_KUBERNETES = "watcher_kubernetes"
 
 
 class InvocationMode(Enum):
@@ -175,10 +179,22 @@ TESTABLE_DBT_RESOURCES = {DbtResourceType.MODEL, DbtResourceType.SOURCE, DbtReso
 DBT_SETUP_ASYNC_TASK_ID = "dbt_setup_async"
 DBT_TEARDOWN_ASYNC_TASK_ID = "dbt_teardown_async"
 
+WATCHER_TASK_WEIGHT_RULE = "absolute"
+CONSUMER_WATCHER_DEFAULT_PRIORITY_WEIGHT = 2
+PRODUCER_WATCHER_DEFAULT_PRIORITY_WEIGHT = 20
 PRODUCER_WATCHER_TASK_ID = "dbt_producer_watcher"
 
-TELEMETRY_URL = "https://astronomer.gateway.scarf.sh/astronomer-cosmos/{telemetry_version}/{cosmos_version}/{airflow_version}/{python_version}/{platform_system}/{platform_machine}/{event_type}/{status}/{dag_hash}/{task_count}/{cosmos_task_count}/{execution_modes}"
-TELEMETRY_VERSION = "v2"
+# Historical telemetry endpoints retained for reference:
+# • v1 (Cosmos 1.8.0–1.10.x)
+#   URL: https://astronomer.gateway.scarf.sh/astronomer-cosmos/{telemetry_version}/{cosmos_version}/{airflow_version}/{python_version}/{platform_system}/{platform_machine}/{event_type}/{status}/{dag_hash}/{task_count}/{cosmos_task_count}
+# • v2 (Cosmos 1.11.0–1.11.x)
+#   URL: https://astronomer.gateway.scarf.sh/astronomer-cosmos/{telemetry_version}/{cosmos_version}/{airflow_version}/{python_version}/{platform_system}/{platform_machine}/{event_type}/{status}/{dag_hash}/{task_count}/{cosmos_task_count}/{execution_modes}
+# • v3 (Cosmos 1.12.0+)
+#   URL: https://astronomer.gateway.scarf.sh/astronomer-cosmos/{telemetry_version}/{event_type}?{query_string}
+TELEMETRY_URL = "https://astronomer.gateway.scarf.sh/astronomer-cosmos/{telemetry_version}/{event_type}?{query_string}"
+TELEMETRY_VERSION = "v3"
 TELEMETRY_TIMEOUT = 1.0
 
 _AIRFLOW3_MAJOR_VERSION = 3
+
+_K8s_WATCHER_MIN_K8S_PROVIDER_VERSION = Version("10.8.0")
