@@ -4,8 +4,9 @@ import inspect
 import logging
 import os
 from abc import ABCMeta, abstractmethod
+from collections.abc import Sequence
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Sequence, Tuple
+from typing import TYPE_CHECKING, Any
 
 import yaml
 from airflow.utils.context import context_merge
@@ -270,7 +271,7 @@ class AbstractDbtBase(metaclass=ABCMeta):
         self,
         context: Context,
         cmd_flags: list[str] | None = None,
-    ) -> Tuple[list[str], dict[str, str | bytes | os.PathLike[Any]]]:
+    ) -> tuple[list[str], dict[str, str | bytes | os.PathLike[Any]]]:
         dbt_cmd = [self.dbt_executable_path]
 
         dbt_cmd.extend(self.dbt_cmd_global_flags)
@@ -318,15 +319,21 @@ class AbstractDbtBase(metaclass=ABCMeta):
 
 
 class DbtBuildMixin:
-    """Mixin for dbt build command."""
+    """
+    Mixin for dbt build command.
+
+    :param full_refresh: whether to add the flag --full-refresh to the dbt build command
+    :param log_format: format for dbt logs (e.g., 'json', 'text'). If provided, adds --log-format flag
+    """
 
     base_cmd = ["build"]
     ui_color = "#8194E0"
 
     template_fields: Sequence[str] = ("full_refresh",)
 
-    def __init__(self, full_refresh: bool | str = False, **kwargs: Any) -> None:
+    def __init__(self, full_refresh: bool | str = False, log_format: str | None = None, **kwargs: Any) -> None:
         self.full_refresh = full_refresh
+        self.log_format = log_format
         super().__init__(**kwargs)
 
     def add_cmd_flags(self) -> list[str]:
@@ -340,6 +347,10 @@ class DbtBuildMixin:
 
         if full_refresh is True:
             flags.append("--full-refresh")
+
+        if self.log_format:
+            flags.append("--log-format")
+            flags.append(self.log_format)
 
         return flags
 
