@@ -3,8 +3,9 @@ from __future__ import annotations
 import inspect
 import re
 from abc import ABC
+from collections.abc import Callable, Sequence
 from os import PathLike
-from typing import TYPE_CHECKING, Any, Callable, Dict, List, Optional, Sequence
+from typing import TYPE_CHECKING, Any
 
 import kubernetes.client as k8s
 from airflow.providers.cncf.kubernetes.backcompat.backwards_compat_converters import (
@@ -200,7 +201,7 @@ class DbtTestWarningHandler(KubernetesPodOperatorCallback):  # type: ignore[misc
         self,
         on_warning_callback: Callable[..., Any],
         operator: KubernetesPodOperator,
-        context: Optional[Context] = None,
+        context: Context | None = None,
     ) -> None:
         self.on_warning_callback = on_warning_callback
         self.operator = operator
@@ -264,7 +265,7 @@ class DbtTestWarningHandler(KubernetesPodOperatorCallback):  # type: ignore[misc
         context_merge(self.context, test_names=test_names, test_results=test_results)
         self.on_warning_callback(self.context)
 
-    def _detect_standard_warnings(self, log_text: str) -> Optional[int]:
+    def _detect_standard_warnings(self, log_text: str) -> int | None:
         """
         Detect warnings using the standard dbt summary pattern.
 
@@ -280,7 +281,7 @@ class DbtTestWarningHandler(KubernetesPodOperatorCallback):  # type: ignore[misc
             return int(match.group(1))
         return None
 
-    def _detect_source_freshness_warnings(self, log_text: str) -> List[Dict[str, Any]]:
+    def _detect_source_freshness_warnings(self, log_text: str) -> list[dict[str, Any]]:
         """
         Detect source freshness warnings from dbt logs.
 
