@@ -6,7 +6,6 @@ from __future__ import annotations
 
 import os
 from dataclasses import dataclass
-from typing import Type
 
 from jinja2 import Environment, FileSystemLoader
 
@@ -22,7 +21,7 @@ class Field:
     airflow_name: str | list[str] | None = None
 
 
-def get_fields_from_mapping(mapping: Type[BaseProfileMapping]) -> list[Field]:
+def get_fields_from_mapping(mapping: type[BaseProfileMapping]) -> list[Field]:
     """
     Generates Field objects from a profile mapping.
     """
