@@ -5,11 +5,13 @@ Astronomer Cosmos is a library for rendering dbt workflows in Airflow.
 
 Contains dags, task groups, and operators.
 """
+
 from __future__ import annotations
 
 from cosmos import settings
 
-__version__ = "1.11.0"
+__version__ = "1.13.0a2"
+
 
 if not settings.enable_memory_optimised_imports:
     from cosmos.airflow.dag import DbtDag
