@@ -1,7 +1,8 @@
 from __future__ import annotations
 
-import importlib
 from copy import deepcopy
+
+from cosmos._utils.importer import load_method_from_module
 
 try:  # Airflow 3
     from airflow.sdk.bases.operator import BaseOperator
@@ -33,8 +34,7 @@ def get_airflow_task(task: Task, dag: DAG, task_group: TaskGroup | None = None) 
     # first, import the operator class from the
     # fully qualified name defined in the task
     module_name, class_name = task.operator_class.rsplit(".", 1)
-    module = importlib.import_module(module_name)
-    Operator = getattr(module, class_name)
+    Operator = load_method_from_module(module_name, class_name)
 
     task_kwargs = task.arguments
     if task.owner != "":
