@@ -1,17 +1,10 @@
 from __future__ import annotations
 
 import os
-import sys
+from functools import cache
 from pathlib import Path
 from unittest.mock import patch
 
-try:
-    from functools import cache
-except ImportError:
-    from functools import lru_cache as cache
-
-
-import airflow
 import pytest
 from airflow.models.dagbag import DagBag
 from airflow.utils.db import create_default_connections
@@ -19,23 +12,20 @@ from airflow.utils.session import provide_session
 from dbt.version import get_installed_version as get_dbt_version
 from packaging.version import Version
 
-from cosmos.constants import PARTIALLY_SUPPORTED_AIRFLOW_VERSIONS
+from cosmos.constants import AIRFLOW_VERSION, PARTIALLY_SUPPORTED_AIRFLOW_VERSIONS
 
 from . import utils as test_utils
 
 EXAMPLE_DAGS_DIR = Path(__file__).parent.parent / "dev/dags"
 AIRFLOW_IGNORE_FILE = EXAMPLE_DAGS_DIR / ".airflowignore"
 DBT_VERSION = Version(get_dbt_version().to_version_string()[1:])
-AIRFLOW_VERSION = Version(airflow.__version__)
-KUBERNETES_DAGS = ["jaffle_shop_kubernetes"]
+KUBERNETES_DAGS = ["jaffle_shop_kubernetes", "jaffle_shop_watcher_kubernetes"]
 
 MIN_VER_DAG_FILE: dict[str, list[str]] = {
-    "2.4": ["cosmos_seed_dag.py"],
     "2.8": ["cosmos_manifest_example.py", "simple_dag_async.py", "cosmos_callback_dag.py"],
 }
 
-IGNORED_DAG_FILES = ["performance_dag.py", "jaffle_shop_kubernetes.py"]
-_PYTHON_VERSION = sys.version_info[:2]
+IGNORED_DAG_FILES = ["performance_dag.py", "jaffle_shop_kubernetes.py", "jaffle_shop_watcher_kubernetes.py"]
 
 # Sort descending based on Versions and convert string to an actual version
 MIN_VER_DAG_FILE_VER: dict[Version, list[str]] = {
@@ -148,11 +138,10 @@ def test_example_dag(session, dag_id: str):
 
 
 @pytest.mark.skipif(
-    _PYTHON_VERSION < (3, 9)
-    or AIRFLOW_VERSION >= Version("3.1.0")  # TODO: Fix https://github.com/astronomer/astronomer-cosmos/issues/2045
+    AIRFLOW_VERSION >= Version("3.1.0")  # TODO: Fix https://github.com/astronomer/astronomer-cosmos/issues/2045
     or AIRFLOW_VERSION < Version("2.8")
     or AIRFLOW_VERSION in PARTIALLY_SUPPORTED_AIRFLOW_VERSIONS,
-    reason="dbt-bigquery only supports Python 3.9 onwards. See PR: https://github.com/apache/airflow/pull/34585 and Airflow 2.9.0 and 2.9.1 have a breaking change in Dataset URIs, and Cosmos errors if `emit_datasets` is not False",
+    reason="Airflow 2.9.0 and 2.9.1 have a breaking change in Dataset URIs (see PR: https://github.com/apache/airflow/pull/34585), and Cosmos errors if `emit_datasets` is not False",
 )
 @patch.dict(
     os.environ,
