@@ -6,12 +6,14 @@ from airflow.models.dagbag import DagBag
 from airflow.utils.db import create_default_connections
 from airflow.utils.session import provide_session
 
+from cosmos.constants import _K8s_WATCHER_MIN_K8S_PROVIDER_VERSION
+
 from . import utils as test_utils
 
 EXAMPLE_DAGS_DIR = Path(__file__).parent.parent / "dev/dags"
 AIRFLOW_IGNORE_FILE = EXAMPLE_DAGS_DIR / ".airflowignore"
 
-KUBERNETES_DAG_FILES = ["jaffle_shop_kubernetes.py"]
+KUBERNETES_DAG_FILES = ["jaffle_shop_kubernetes.py", "jaffle_shop_watcher_kubernetes.py"]
 
 
 @provide_session
@@ -39,6 +41,24 @@ def get_all_dag_files():
 def test_example_dag_kubernetes(session):
     get_all_dag_files()
     db = DagBag(EXAMPLE_DAGS_DIR, include_examples=False)
-    # for dag_id in KUBERNETES_DAG_FILES:
+    assert not db.import_errors
     dag = db.get_dag("jaffle_shop_kubernetes")
+    test_utils.run_dag(dag)
+
+
+from airflow.providers.cncf.kubernetes import __version__ as airflow_k8s_provider_version
+from packaging.version import Version
+
+
+@pytest.mark.skipif(
+    Version(airflow_k8s_provider_version) < _K8s_WATCHER_MIN_K8S_PROVIDER_VERSION,
+    reason="This feature is only available for K8s provider 10.8.0 and above",
+)
+@pytest.mark.integration
+def test_example_dag_watcher_kubernetes(session):
+    get_all_dag_files()
+    db = DagBag(EXAMPLE_DAGS_DIR, include_examples=False)
+    dag = db.get_dag("jaffle_shop_watcher_kubernetes")
+    assert not db.import_errors
+    assert dag is not None
     test_utils.run_dag(dag)
