@@ -1,3 +1,4 @@
+import os
 from pathlib import Path
 from unittest.mock import MagicMock, patch
 
@@ -76,15 +77,23 @@ def test_configure_remote_target_path_no_remote_target():
 
 
 def test_construct_dest_file_path(dummy_kwargs):
-    """Test _construct_dest_file_path."""
+    """Test construct_dest_file_path."""
     dest_target_dir = Path("/dest")
     source_target_dir = Path("/project_dir/target")
     file_path = "/project_dir/target/subdir/file.txt"
 
+    rel_path = os.path.relpath(file_path, source_target_dir)
+    context = dummy_kwargs["context"]
+    task_run_identifier = (
+        f"{context['dag'].dag_id}"
+        f"/{context['run_id']}"
+        f"/{context['task_instance'].task_id}"
+        f"/{context['task_instance'].try_number}"
+    )
+
     expected_path = "/dest/test_dag/test_run_id/test_task/1/target/subdir/file.txt"
     assert (
-        _construct_dest_file_path(dest_target_dir, file_path, source_target_dir, DEFAULT_TARGET_PATH, **dummy_kwargs)
-        == expected_path
+        _construct_dest_file_path(dest_target_dir, rel_path, task_run_identifier, DEFAULT_TARGET_PATH) == expected_path
     )
 
 
@@ -98,12 +107,14 @@ def test_upload_artifacts_to_cloud_storage_no_remote_path():
 @pytest.mark.skipif(not AIRFLOW_IO_AVAILABLE, reason="Airflow did not have Object Storage until the 2.8 release")
 def test_upload_artifacts_to_cloud_storage_success(dummy_kwargs):
     """Test upload_artifacts_to_cloud_storage with valid setup."""
-    with patch(
-        "cosmos.io._configure_remote_target_path",
-        return_value=(Path("/dest"), "conn_id"),
-    ) as mock_configure, patch("pathlib.Path.rglob") as mock_rglob, patch(
-        "cosmos.io.ObjectStoragePath.copy"
-    ) as mock_copy:
+    with (
+        patch(
+            "cosmos.io._configure_remote_target_path",
+            return_value=(Path("/dest"), "conn_id"),
+        ) as mock_configure,
+        patch("pathlib.Path.rglob") as mock_rglob,
+        patch("cosmos.io.ObjectStoragePath.copy") as mock_copy,
+    ):
         mock_file1 = MagicMock(spec=Path)
         mock_file1.is_file.return_value = True
         mock_file1.__str__.return_value = "/project_dir/target/file1.txt"
