@@ -14,30 +14,26 @@ def test_enable_cache_env_var():
 
 
 def test_enable_memory_optimised_imports_true(monkeypatch):
-    script = textwrap.dedent(
-        """
+    script = textwrap.dedent("""
             import os
             os.environ["AIRFLOW__COSMOS__ENABLE_MEMORY_OPTIMISED_IMPORTS"] = "True"
             import cosmos
             assert cosmos.settings.enable_memory_optimised_imports is True
             assert not hasattr(cosmos, "DbtDag")
-        """
-    )
+        """)
 
     result = subprocess.run(["python", "-c", script], capture_output=True, text=True)
     assert result.returncode == 0, result.stderr
 
 
 def test_enable_memory_optimised_imports_false(monkeypatch):
-    script = textwrap.dedent(
-        """
+    script = textwrap.dedent("""
             import os
             os.environ["AIRFLOW__COSMOS__ENABLE_MEMORY_OPTIMISED_IMPORTS"] = "False"
             import cosmos
             assert cosmos.settings.enable_memory_optimised_imports is False
             assert hasattr(cosmos, "DbtDag")
-        """
-    )
+        """)
 
     result = subprocess.run(["python", "-c", script], capture_output=True, text=True)
     assert result.returncode == 0, result.stderr
