@@ -7,6 +7,45 @@ import pytest
 from cosmos import telemetry
 
 
+def test_compress_telemetry_metadata_is_deterministic():
+    """Test that compressing the same metadata multiple times produces identical output."""
+    metadata = {
+        "cosmos_version": "1.8.0",
+        "airflow_version": "2.10.1",
+        "python_version": "3.11",
+        "execution_mode": "local",
+        "install_deps": True,
+    }
+
+    # Compress the same metadata multiple times
+    compressed_1 = telemetry._compress_telemetry_metadata(metadata)
+    compressed_2 = telemetry._compress_telemetry_metadata(metadata)
+    compressed_3 = telemetry._compress_telemetry_metadata(metadata)
+
+    # All compressed outputs should be identical (deterministic)
+    assert compressed_1 == compressed_2
+    assert compressed_2 == compressed_3
+
+
+def test_compress_decompress_telemetry_metadata_roundtrip():
+    """Test that metadata can be compressed and decompressed correctly."""
+    original_metadata = {
+        "cosmos_version": "1.8.0",
+        "airflow_version": "2.10.1",
+        "python_version": "3.11",
+        "execution_mode": "local",
+        "install_deps": False,
+        "dbt_command": "run",
+    }
+
+    # Compress and decompress
+    compressed = telemetry._compress_telemetry_metadata(original_metadata)
+    decompressed = telemetry._decompress_telemetry_metadata(compressed)
+
+    # Should get back the original metadata
+    assert decompressed == original_metadata
+
+
 def test_should_emit_is_true_by_default():
     assert telemetry.should_emit()
 
@@ -34,7 +73,6 @@ def test_collect_standard_usage_metrics():
         "platform_machine",
         "platform_system",
         "python_version",
-        "variables",
     ]
     assert sorted(metrics.keys()) == expected_keys
 
@@ -62,13 +100,13 @@ def test_emit_usage_metrics_is_unsuccessful(mock_httpx_get, caplog):
     }
     is_success = telemetry.emit_usage_metrics(sample_metrics)
     mock_httpx_get.assert_called_once_with(
-        f"""https://astronomer.gateway.scarf.sh/astronomer-cosmos/v2/1.8.0a4/2.10.1/3.11/darwin/amd64/dag_run/success/d151d1fa2f03270ea116cc7494f2c591/3/3/local""",
+        f"""https://astronomer.gateway.scarf.sh/astronomer-cosmos/v3/dag_run?cosmos_version=1.8.0a4&airflow_version=2.10.1&python_version=3.11&platform_system=darwin&platform_machine=amd64&status=success&dag_hash=d151d1fa2f03270ea116cc7494f2c591&task_count=3&cosmos_task_count=3&execution_modes=local""",
         timeout=1.0,
         follow_redirects=True,
     )
     assert not is_success
-    log_msg = f"""Unable to emit usage metrics to https://astronomer.gateway.scarf.sh/astronomer-cosmos/v2/1.8.0a4/2.10.1/3.11/darwin/amd64/dag_run/success/d151d1fa2f03270ea116cc7494f2c591/3/3/local. Status code: 404. Message: Non existent URL"""
-    assert caplog.text.startswith("WARNING")
+    log_msg = f"""Unable to emit usage metrics to https://astronomer.gateway.scarf.sh/astronomer-cosmos/v3/dag_run?cosmos_version=1.8.0a4&airflow_version=2.10.1&python_version=3.11&platform_system=darwin&platform_machine=amd64&status=success&dag_hash=d151d1fa2f03270ea116cc7494f2c591&task_count=3&cosmos_task_count=3&execution_modes=local. Status code: 404. Message: Non existent URL"""
+    assert "WARNING" in caplog.text
     assert log_msg in caplog.text
 
 
@@ -89,17 +127,18 @@ def test_emit_usage_metrics_fails(mock_httpx_get, caplog):
     }
     is_success = telemetry.emit_usage_metrics(sample_metrics)
     mock_httpx_get.assert_called_once_with(
-        f"""https://astronomer.gateway.scarf.sh/astronomer-cosmos/v2/1.8.0a4/2.10.1/3.11/darwin/amd64/dag_run/success/d151d1fa2f03270ea116cc7494f2c591/3/3/local""",
+        f"""https://astronomer.gateway.scarf.sh/astronomer-cosmos/v3/dag_run?cosmos_version=1.8.0a4&airflow_version=2.10.1&python_version=3.11&platform_system=darwin&platform_machine=amd64&status=success&dag_hash=d151d1fa2f03270ea116cc7494f2c591&task_count=3&cosmos_task_count=3&execution_modes=local""",
         timeout=1.0,
         follow_redirects=True,
     )
     assert not is_success
-    log_msg = f"""Unable to emit usage metrics to https://astronomer.gateway.scarf.sh/astronomer-cosmos/v2/1.8.0a4/2.10.1/3.11/darwin/amd64/dag_run/success/d151d1fa2f03270ea116cc7494f2c591/3/3/local. An HTTPX connection error occurred: Something is not right."""
-    assert caplog.text.startswith("WARNING")
+    log_msg = f"""Unable to emit usage metrics to https://astronomer.gateway.scarf.sh/astronomer-cosmos/v3/dag_run?cosmos_version=1.8.0a4&airflow_version=2.10.1&python_version=3.11&platform_system=darwin&platform_machine=amd64&status=success&dag_hash=d151d1fa2f03270ea116cc7494f2c591&task_count=3&cosmos_task_count=3&execution_modes=local. An HTTPX connection error occurred: Something is not right."""
+    assert "WARNING" in caplog.text
     assert log_msg in caplog.text
 
 
 @pytest.mark.integration
+@pytest.mark.flaky(reruns=2, reruns_delay=1)
 def test_emit_usage_metrics_succeeds(caplog):
     caplog.set_level(logging.DEBUG)
     sample_metrics = {
@@ -118,19 +157,19 @@ def test_emit_usage_metrics_succeeds(caplog):
     is_success = telemetry.emit_usage_metrics(sample_metrics)
     assert is_success
     assert caplog.text.startswith("DEBUG")
-    assert "Telemetry is enabled. Emitting the following usage metrics to" in caplog.text
+    assert "Telemetry is enabled. Emitting the following usage metrics for event type" in caplog.text
 
 
 @patch("cosmos.telemetry.should_emit", return_value=False)
 def test_emit_usage_metrics_if_enabled_fails(mock_should_emit, caplog):
     caplog.set_level(logging.DEBUG)
     assert not telemetry.emit_usage_metrics_if_enabled("any", {})
-    assert caplog.text.startswith("DEBUG")
+    assert "DEBUG" in caplog.text
     assert "Telemetry is disabled. To enable it, export AIRFLOW__COSMOS__ENABLE_TELEMETRY=True." in caplog.text
 
 
 @patch("cosmos.telemetry.should_emit", return_value=True)
-@patch("cosmos.telemetry.collect_standard_usage_metrics", return_value={"k1": "v1", "k2": "v2", "variables": {}})
+@patch("cosmos.telemetry.collect_standard_usage_metrics", return_value={"k1": "v1", "k2": "v2"})
 @patch("cosmos.telemetry.emit_usage_metrics")
 def test_emit_usage_metrics_if_enabled_succeeds(
     mock_emit_usage_metrics, mock_collect_standard_usage_metrics, mock_should_emit
@@ -141,5 +180,4 @@ def test_emit_usage_metrics_if_enabled_succeeds(
         "k1": "v1",
         "k2": "v2",
         "event_type": "any",
-        "variables": {"k2": "v2"},
     }
